Restrict workflow to main branch
The check and test jobs run just check and just test respectively, which
we can run locally. Should we forget, pre-commit will run the same tools
for us. We want to add a tag-new-version job, which should run after
check and test on push to main. To make adding that job easier, and
without loosing checking and testing, we restrict this workflow to the
main branch.