Add --output-dir argument
Adding this argument allows us to pass the output dataframes, or the
cohorts produced by cohort-joiner, to the measures framework. We also
remove the ability to write the output dataframes to the same directory
as the input dataframes, or the cohorts produced by cohort-extractor,
because this makes it hard to construct an audit trail.